requirements.txt
numpy
scikit-learn
torch
torchvision
pydicom
Pillow
